+++ internal/report/report.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	"github.com/sebastienrousseau/scout"
-	"github.com/sebastienrousseau/scout/attestation"
+	"github.com/sebastienrousseau/scout-reporting/attestation"
 	"github.com/sebastienrousseau/scout/internal/probe"
 	"github.com/sebastienrousseau/scout/internal/telemetry"
 )